chore: abstracted loading from DB and cleaned up script
